rename errorHandler to requestHandler
